refactor: rename the linked-dependencies component
