+++ scopes/dependencies/linked-dependencies/index.ts
