+++ scopes/dependencies/linked-dependencies/linked-dependencies.docs.mdx
@@ -6,7 +6,7 @@ description: 'Links linked dependencies to node_modules'
 Usage:
 
 ```ts
-import { createLinks } from '@teambit/toolbox.fs.linked-dependencies';
+import { createLinks } from '@teambit/dependencies.linked-dependencies';
 
 await createLinks('/Users/Foo/src/project', {
   lodash: 'link:../lodash-fork',
