+++ scopes/dependencies/linked-dependencies/linked-dependencies.ts
